+++ cpp/pmft/PMFTXY2D.h
@@ -1,6 +1,9 @@
 // Copyright (c) 2010-2018 The Regents of the University of Michigan
 // This file is part of the freud project, released under the BSD 3-Clause License.
 
+#ifndef PMFTXY2D_H
+#define PMFTXY2D_H
+
 #include <memory>
 #include <ostream>
 #include <tbb/tbb.h>
@@ -10,9 +13,6 @@
 #include "LinkCell.h"
 #include "Index1D.h"
 #include "PMFT.h"
-
-#ifndef _PMFTXY2D_H__
-#define _PMFTXY2D_H__
 
 /*! \internal
     \file PMFTXY2D.h
@@ -40,7 +40,7 @@ class PMFTXY2D : public PMFT
         PMFTXY2D(float max_x, float max_y, unsigned int n_bins_x, unsigned int n_bins_y);
 
         //! Reset the PCF array to all zeros
-        virtual void resetPCF();
+        virtual void reset();
 
         /*! Compute the PCF for the passed in set of points. The result will
          *  be added to previous values of the PCF.
@@ -101,4 +101,4 @@ class PMFTXY2D : public PMFT
 
 }; }; // end namespace freud::pmft
 
-#endif // _PMFTXY2D_H__
+#endif // PMFTXY2D_H