+++ cpp/pmft/PMFTXYZ.h
@@ -1,6 +1,9 @@
 // Copyright (c) 2010-2018 The Regents of the University of Michigan
 // This file is part of the freud project, released under the BSD 3-Clause License.
 
+#ifndef PMFTXYZ_H
+#define PMFTXYZ_H
+
 #include <memory>
 #include <ostream>
 #include <tbb/tbb.h>
@@ -10,9 +13,6 @@
 #include "LinkCell.h"
 #include "Index1D.h"
 #include "PMFT.h"
-
-#ifndef _PMFTXYZ_H__
-#define _PMFTXYZ_H__
 
 /*! \internal
     \file PMFTXYZ.h
@@ -38,7 +38,7 @@ class PMFTXYZ : public PMFT
         PMFTXYZ(float max_x, float max_y, float max_z, unsigned int n_bins_x, unsigned int n_bins_y, unsigned int n_bins_z, vec3<float> shiftvec);
 
         //! Reset the PCF array to all zeros
-        virtual void resetPCF();
+        virtual void reset();
 
         /*! Compute the PCF for the passed in set of points. The function will be added to previous values
             of the pcf
@@ -118,4 +118,4 @@ class PMFTXYZ : public PMFT
 
 }; }; // end namespace freud::pmft
 
-#endif // _PMFTXYZ_H__
+#endif // PMFTXYZ_H