+++ freud/__init__.py
@@ -20,7 +20,9 @@ from . import parallel
 from . import box
 from . import bond
 from . import cluster
+from . import common
 from . import density
+from . import environment
 from . import index
 from . import interface
 from . import kspace
@@ -28,6 +30,5 @@ from . import locality
 from . import order
 from . import pmft
 from . import voronoi
-from . import common
 
 __version__ = '0.8.2'